Add defaultRouteToVpc and extraContainerRoutes opt
* defaultRouteToVpc allows you to set the default route in the container
namespace to the VPC router. This is useful for AWS networks that
*don't* assign public IPs to their instances and rely on NAT boxes
listed in the VPC route table. Or if you happen to have other networks
reachable from the VPC route table.

* extraContainerRoutes allows you to add extra routes in the container
that route back to the host via the veth pair. This is useful if you
have things listening on the host namespace but have updated the default
route to go via the VPC router (as above).